+++ packages/experiment-browser/src/index.ts
@@ -36,3 +36,8 @@ export {
 } from './types/plugin';
 export { Logger, LogLevel } from './types/logger';
 export { ConsoleLogger } from './logger/consoleLogger';
+export {
+  EvaluationConditionResult,
+  EvaluationSegmentResult,
+  FlagEvaluationTrace,
+} from '@amplitude/experiment-core';